+++ packages/ui/src/components/custom/icon.tsx
@@ -1,5 +1,4 @@
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription } from "../ui/dialog";
-import { DialogTrigger } from "../ui/dialog";
+import { IconWithDownloadDialog } from "./app-download-dialog";
 
 type IconProps = {
   className?: string;
@@ -41,67 +40,6 @@ export function ArticleLikeIcon({ className = 'size-5' }: IconProps) {
         d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
       />
     </svg>
-  );
-}
-
-// アプリダウンロードダイアログコンポーネントを作成
-export function AppDownloadDialog() {
-  return (
-    <Dialog>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>アプリをダウンロード</DialogTitle>
-          <DialogDescription>
-            より良い体験のために、アプリをダウンロードしてください。
-          </DialogDescription>
-        </DialogHeader>
-        <div className="flex flex-col gap-4 py-4">
-          <p>アプリを使うとより快適に利用できます。今すぐダウンロードしましょう！</p>
-          <a 
-            href="/download-app" 
-            className="flex items-center justify-center rounded-md bg-primary px-4 py-2 text-white"
-          >
-            アプリをダウンロード
-          </a>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-// アイコンをダイアログトリガーとして使用するラッパーコンポーネント
-export function IconWithDownloadDialog({ 
-  children, 
-  className 
-}: { 
-  children: React.ReactNode; 
-  className?: string 
-}) {
-  return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <button className={className}>
-          {children}
-        </button>
-      </DialogTrigger>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>アプリをダウンロード</DialogTitle>
-          <DialogDescription>
-            より良い体験のために、アプリをダウンロードしてください。
-          </DialogDescription>
-        </DialogHeader>
-        <div className="flex flex-col gap-4 py-4">
-          <p>アプリを使うとより快適に利用できます。今すぐダウンロードしましょう！</p>
-          <a 
-            href="/download-app" 
-            className="flex items-center justify-center rounded-md bg-primary px-4 py-2 text-white"
-          >
-            アプリをダウンロード
-          </a>
-        </div>
-      </DialogContent>
-    </Dialog>
   );
 }
