+++ packages/ui/src/components/custom/store-badge.tsx
@@ -2,6 +2,7 @@
 
 import Image from "next/image";
 import Link from "next/link";
+import React from "react";
 
 type StoreBadgeProps = {
   store: "apple" | "google";
